[Parsing] -> [Master] | Parsing complete. Map dump done. Map validation done (except for path validations)

+++ includes/farm.h
@@ -6,7 +6,7 @@
 /*   By: cpoulain <cpoulain@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/08/13 12:26:43 by cpoulain          #+#    #+#             */
-/*   Updated: 2025/08/13 12:27:46 by cpoulain         ###   ########.fr       */
+/*   Updated: 2025/08/13 16:09:37 by cpoulain         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,9 +16,9 @@
 
 typedef	struct	s_farm
 {
-	int		ant_count;
+	size_t	ant_count;
 	t_room	**rooms;
-	int		room_count;
+	size_t	room_count;
 	t_room	*start;
 	t_room	*end;
 }	t_farm;
