+++ includes/room.h
@@ -6,11 +6,13 @@
 /*   By: cpoulain <cpoulain@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/08/13 12:23:34 by cpoulain          #+#    #+#             */
-/*   Updated: 2025/08/13 12:26:39 by cpoulain         ###   ########.fr       */
+/*   Updated: 2025/08/14 16:12:53 by cpoulain         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #pragma once
+
+#include <stddef.h>
 
 typedef enum e_room_type
 {
@@ -27,5 +29,5 @@ typedef struct s_room
 	int			id;
 	t_room_type	type;
 	struct		s_room	**links;
-	int			link_count;
+	size_t		link_count;
 }	t_room;
